loadInvaderWave
Added a function to read from a file to create a wave of invaders, a
wave just being an array because we have a hard bound on the size of a
wave. Why bring linked lists in when we don't have to?

It compiled the first time, so that means it must have a shit ton of
segfaults once I actually start testing it.